+++ quartz-js/components/scope.js
@@ -17,7 +17,8 @@ if (typeof(vhxm) !== 'undefined') {
     },
     select: {
       ui: {
-        list: {}
+        list: {},
+        media_item: {}
       }
     },
     autosuggest: {